Fix issue with advanced search to pick the correct enrollment
As the current implementation the devices returned from the search query are filtered by device id which causes invalid devices to return. This commit fixes the issue by filtering the devices by enrollment id

Also all the devices should be shown to users which has the permission to view any device. This commit checks the permission of the users and return owners device if user is not a privileged.

Fixes entgra/product-iots#13
Fixes entgra/product-iots#14